+++ packages/api/api.ts
@@ -30,7 +30,7 @@ export interface UnchainedServerOptions {
   rolesOptions: any;
 }
 
-const UNCHAINED_API_VERSION = '0.61.0'; // eslint-disable-line
+const UNCHAINED_API_VERSION = '0.61.1'; // eslint-disable-line
 
 export const createContextResolver = (unchained: UnchainedAPI) => async ({
   req,