+++ packages/core-accountsjs/settings.js
@@ -1,10 +1,8 @@
 import { accountsPassword } from './accounts-password';
 
-const sendVerificationEmail = true;
-
 const settings = {
-  load({ sendVerificationEmailAfterSignup = sendVerificationEmail } = {}) {
-    accountsPassword.options.sendVerificationEmailAfterSignup = sendVerificationEmailAfterSignup;
+  load() {
+    accountsPassword.options.sendVerificationEmailAfterSignup = false;
   },
 };