+++ packages/core-users/settings.js
@@ -1,0 +1,13 @@
+const defaultAutoMessagingAfterUserCreation = true;
+
+const settings = {
+  autoMessagingAfterUserCreation: null,
+  load({
+    autoMessagingAfterUserCreation = defaultAutoMessagingAfterUserCreation,
+  } = {}) {
+    this.autoMessagingAfterUserCreation =
+      autoMessagingAfterUserCreation ?? defaultAutoMessagingAfterUserCreation;
+  },
+};
+
+export default settings;